requirements.txt
PyQt5
scapy
mss
threading
thread
urllib
socket
hashlib
fernet
zipfile
pexpect
ftplib
http
requests
math
sys
time
distutils
subprocess
shutil
pynput
jason
struct
string
base64
 